address an issue with code for a product filter.
And, the issue propagates into one unit test.
Code using this must change. apigee-internal/cli/lib/gateway.js is a consumer that must alter parameters to calls.

Previously, the env field was not passed to methods in network.js.
Hence, the filter code added elsewhere did not function and was reported by a customer.
However, the filter code was not in error itself, only the absence of the parameter caused the issue.